+++ android-sdk-ui/src/com/appboy/push/AppboyNotificationActionUtils.java
@@ -1,13 +1,9 @@
 package com.appboy.push;
 
-import android.Manifest;
 import android.annotation.TargetApi;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.Bitmap;
-import android.net.Uri;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.NotificationCompat;
@@ -16,17 +12,12 @@ import com.appboy.Appboy;
 import com.appboy.AppboyGcmReceiver;
 import com.appboy.Constants;
 import com.appboy.configuration.AppboyConfigurationProvider;
-import com.appboy.support.AppboyImageUtils;
 import com.appboy.support.AppboyLogger;
 import com.appboy.support.IntentUtils;
-import com.appboy.support.PermissionUtils;
 import com.appboy.support.StringUtils;
 
 public class AppboyNotificationActionUtils {
   private static final String TAG = String.format("%s.%s", Constants.APPBOY_LOG_TAG_PREFIX, AppboyNotificationActionUtils.class.getName());
-  public static final String DEFAULT_LOCAL_STORAGE_FOLDER = "Shared Photos";
-  public static final String IMAGE_MIME_TYPE = "image/*";
-  public static final String TEXT_MIME_TYPE = "text/plain";
 
   /**
    * Add notification actions to the provided notification builder.
@@ -35,7 +26,7 @@ public class AppboyNotificationActionUtils {
    *
    * “ab_a*_id”: action button id, used for analytics - optional
    * “ab_a*_t”: action button text - optional
-   * “ab_a*_a”: action type, one of “ab_uri”, ”ab_none”, “ab_share”, “ab_open” (open the app), or a custom defined action - required
+   * “ab_a*_a”: action type, one of “ab_uri”, ”ab_none”, “ab_open” (open the app) - required
    * “ab_a*_uri”: uri, only used when the action is “uri” - required only when action is “uri”
    *
    * The * is replaced with an integer string depending on the button being described
@@ -73,8 +64,6 @@ public class AppboyNotificationActionUtils {
    * receiver when an Appboy notification action button is clicked. The GCM/ADM receiver passes on
    * the intent from the notification action button click intent.
    *
-   * See {@link #logNotificationActionClicked} and {@link #handleShareActionClicked}
-   *
    * @param context
    * @param intent the action button click intent
    */
@@ -89,10 +78,9 @@ public class AppboyNotificationActionUtils {
       int notificationId = intent.getIntExtra(Constants.APPBOY_PUSH_NOTIFICATION_ID, Constants.APPBOY_DEFAULT_NOTIFICATION_ID);
 
       // Logs that the notification action was clicked.
-      // Notification actions with no click actions are not logged.
-      if (!actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_NONE)) {
-        logNotificationActionClicked(context, intent);
-      }
+      // Click analytics for all action types are logged.
+      logNotificationActionClicked(context, intent);
+
       if (actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_URI) || actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_OPEN)) {
         AppboyNotificationUtils.cancelNotification(context, notificationId);
         context.sendBroadcast(new Intent(Intent.ACTION_CLOSE_SYSTEM_DIALOGS));
@@ -109,52 +97,14 @@ public class AppboyNotificationActionUtils {
         if (appConfigurationProvider.getHandlePushDeepLinksAutomatically()) {
           AppboyNotificationUtils.routeUserWithNotificationOpenedIntent(context, intent);
         }
-      } else if (actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_SHARE)) {
-        AppboyNotificationUtils.cancelNotification(context, notificationId);
-        context.sendBroadcast(new Intent(Intent.ACTION_CLOSE_SYSTEM_DIALOGS));
-        handleShareActionClicked(context, intent);
       } else if (actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_NONE)) {
         AppboyNotificationUtils.cancelNotification(context, notificationId);
       } else {
-        AppboyLogger.i(TAG, "Custom notification action button clicked. Doing nothing and passing on data to client receiver.");
-        AppboyNotificationUtils.sendNotificationOpenedBroadcast(context, intent);
+        AppboyLogger.w(TAG, "Unknown notification action button clicked. Doing nothing.");
       }
     } catch (Exception e) {
       AppboyLogger.e(TAG, "Caught exception while handling notification action button click.", e);
     }
-  }
-
-  /**
-   * Whether the preconditions for sharing an image are met.
-   *
-   * These include an image URI in the push extras, external write permission, and whether a
-   * downloaded image is available for the current push notification.
-   *
-   * @param context
-   * @param appboyExtras
-   * @return whether sharing an image is currently possible.
-   */
-  static boolean canShareImage(Context context, Bundle appboyExtras) {
-    return appboyExtras != null
-        && appboyExtras.containsKey(Constants.APPBOY_PUSH_BIG_IMAGE_URL_KEY)
-        && PermissionUtils.hasPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE);
-  }
-
-  /**
-   * @param actionType
-   * @return true if the given action type is not an Appboy preset action type
-   */
-  static boolean isCustomActionType(String actionType) {
-    if (actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_URI)) {
-      return false;
-    } else if (actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_OPEN)) {
-      return false;
-    } else if (actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_NONE)) {
-      return false;
-    } else if (actionType.equals(Constants.APPBOY_PUSH_ACTION_TYPE_SHARE)) {
-      return false;
-    }
-    return true;
   }
 
   /**
@@ -174,7 +124,6 @@ public class AppboyNotificationActionUtils {
     notificationActionExtras.putString(Constants.APPBOY_ACTION_TYPE_KEY, actionType);
     notificationActionExtras.putString(Constants.APPBOY_ACTION_ID_KEY, getActionFieldAtIndex(actionIndex, notificationExtras, Constants.APPBOY_PUSH_ACTION_ID_KEY_TEMPLATE));
     notificationActionExtras.putString(Constants.APPBOY_ACTION_URI_KEY, getActionFieldAtIndex(actionIndex, notificationExtras, Constants.APPBOY_PUSH_ACTION_URI_KEY_TEMPLATE));
-    notificationActionExtras.putBoolean(Constants.APPBOY_ACTION_IS_CUSTOM_ACTION_KEY, isCustomActionType(actionType));
 
     Intent sendIntent = new Intent(Constants.APPBOY_ACTION_CLICKED_ACTION).setClass(context, AppboyNotificationUtils.getNotificationReceiverClass());
     sendIntent.putExtras(notificationActionExtras);
@@ -207,88 +156,6 @@ public class AppboyNotificationActionUtils {
     AppboyLogger.i(TAG, "Logging push action click to Appboy. Campaign Id: " + campaignId + " Action Button Id: " + actionButtonId);
     Appboy.getInstance(context).logPushNotificationActionClicked(campaignId, actionButtonId);
 
-  }
-
-  /**
-   * Handles sharing the action button click intent's contents to apps that can handle
-   * the ACTION_SEND intent.
-   *
-   * @param context
-   * @param intent the action button click intent
-   */
-  private static void handleShareActionClicked(Context context, Intent intent) {
-    new ShareTask(context).execute(intent);
-  }
-
-  /**
-   * Utility class to perform share intent creation on a background thread.
-   *
-   * Using a background thread is required since when sharing an image, the full image will be
-   * re-downloaded, which must take place on a background thread.
-   */
-  private static class ShareTask extends AsyncTask<Intent, Integer, Intent> {
-    private final Context mContext;
-
-    public ShareTask(Context context) {
-      mContext = context;
-    }
-
-    @Override
-    protected Intent doInBackground(Intent... intents) {
-      if (mContext != null) {
-        return createShareActionIntent(mContext, intents[0]);
-      } else {
-        return null;
-      }
-    }
-
-    @Override
-    protected void onPostExecute(Intent shareIntent) {
-      if (mContext != null) {
-        if (shareIntent != null) {
-          mContext.startActivity(shareIntent);
-        } else {
-          AppboyLogger.w(TAG, "Null share intent received. Not starting share intent.");
-        }
-      }
-    }
-  }
-
-  /**
-   * Creates an intent that can used to share the contents of the notification intent passed in.
-   *
-   * Handles downloading and storing any push images locally, which is required for sharing them.
-   *
-   * Must be executed on a background thread.
-   *
-   * @param context
-   * @param intent the action button click intent
-   * @return a shareable intent
-   */
-  private static Intent createShareActionIntent(Context context, Intent intent) {
-    Bundle notificationExtras = intent.getExtras();
-    Intent shareIntent = new Intent(Intent.ACTION_SEND);
-    shareIntent.putExtra(Intent.EXTRA_SUBJECT, notificationExtras.getString(Constants.APPBOY_PUSH_TITLE_KEY));
-    shareIntent.putExtra(Intent.EXTRA_TEXT, notificationExtras.getString(Constants.APPBOY_PUSH_CONTENT_KEY));
-
-    Bundle appboyExtras = notificationExtras.getBundle(Constants.APPBOY_PUSH_EXTRAS_KEY);
-    if (canShareImage(context, appboyExtras)) {
-      String fileName = Long.toString(System.currentTimeMillis());
-
-      String imageUrl = appboyExtras.getString(Constants.APPBOY_PUSH_BIG_IMAGE_URL_KEY);
-      Bitmap imageBitmap = AppboyImageUtils.getBitmap(Uri.parse(imageUrl));
-
-      Uri localImageUri = AppboyImageUtils
-          .storePushBitmapInExternalStorage(context.getApplicationContext(), imageBitmap, fileName, DEFAULT_LOCAL_STORAGE_FOLDER);
-
-      shareIntent.setType(IMAGE_MIME_TYPE);
-      shareIntent.putExtra(Intent.EXTRA_STREAM, localImageUri);
-      shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-    } else {
-      shareIntent.setType(TEXT_MIME_TYPE);
-    }
-    shareIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-    return shareIntent;
   }
 
   /**